feat(auth, profile): реалізація ендпоінтів та рефакторинг стилів
