--- models/User.mjs
@@ -1,11 +1,0 @@
-import mongoose from 'mongoose';
-
-const userSchema = new mongoose.Schema(
-    {
-        username: { type: String, required: true, unique: true, trim: true },
-        password: { type: String, required: true },
-    },
-    { timestamps: true }
-);
-
-export default mongoose.model('User', userSchema);
